bus API/requirements.txt
fastapi
uvicorn
sqlalchemy
pymysql
python-dotenv
requests
cryptography
